[調整]入力のバグが発覚したため修正(勝手に入力される状態になっていた)

+++ InputCheckerTest/DirectX12 MainProject/_Classes/_GeneralPurpose/_InputClasses/GamePadBase.h
@@ -24,7 +24,7 @@ public:
 	virtual void Accepts(int index) = 0;
 
 	inline bool OnOffKey(PAD inputKey) {
-		return (flag_ & (1 << inputKey));
+		return (flag_ & (1 << (int)inputKey));
 	}
 
 	inline bool AnyKey() {
